+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Security: Fix arbitrary Sieve script injection via a filter rule name bypassing `managesieve_disabled_actions`
 - Security: Fix RCE via `cmd_learn` driver of markasjunk plugin
 - Security: Fix IMAP command injection via mail search and LITERAL+ byte-count desynchronization
+- Security: Fix password's modoboa driver leak of an authentication token to a user-controlled host
 
 ## Release 1.6.17